# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'Links API'
  description: ...
  version: 1.0.0
paths:
...
  /user:
    get:
      tags:
        - User
      summary: 'Get paginated list of users.'
      description: 'Only for admins.'
      operationId: d8640887f6a3f40193deeb33300309da
      responses:
        200:
          description: 'Paginated list.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsersListPaginated'
        401:
          description: Unauthenticated.
        403:
          description: 'Permissions error.'
      security:
        -
          sanctum: []
    post:
      tags:
        - User
      summary: 'Create new user manually.'
      description: 'Only for admins.'
      operationId: a2024e7428c2b58c64a6da90003a7240
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StoreUserRequest'
      responses:
        200:
          description: 'Paginated list.'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Info message.'
                  entity:
                    $ref: '#/components/schemas/User'
                type: object
        401:
          description: Unauthenticated.
        403:
          description: 'Permissions error.'
        422:
          description: 'Validation error.'
      security:
        -
          sanctum: []
  '/user/{id}':
    get:
      tags:
        - User
      summary: 'Show user by id.'
      description: 'Only for admins.'
      operationId: 550c0be2ab85ba78b49fb11dfa03e192
      parameters:
        -
          name: id
          in: path
          description: 'ID of user.'
          required: true
          example: '1'
      responses:
        200:
          description: 'User object.'
          content:
            application/json:
              schema:
                properties:
                  entity:
                    $ref: '#/components/schemas/User'
                type: object
        401:
          description: Unauthenticated.
        403:
          description: 'Permissions error.'
        404:
          description: 'User not found.'
      security:
        -
          sanctum: []
    put:
      tags:
        - User
      summary: 'Update all user data.'
      description: 'Only for admins.'
      operationId: 9fb54e5f5daf60dc628e49da7ba764d5
      parameters:
        -
          name: id
          in: path
          description: 'ID of user.'
          required: true
          example: '1'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PutUserRequest'
      responses:
        200:
          description: 'User updated.'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Info message.'
                  entity:
                    $ref: '#/components/schemas/User'
                  old_entity:
                    $ref: '#/components/schemas/User'
                type: object
        401:
          description: Unauthenticated.
        403:
          description: 'Permissions error.'
        404:
          description: 'User not found.'
        422:
          description: 'Validation error.'
      security:
        -
          sanctum: []
    delete:
      tags:
        - User
      summary: 'Delete user.'
      description: 'Only for admins.'
      operationId: eca0ff485f5a960a804f0992ae1ec9be
      parameters:
        -
          name: id
          in: path
          description: 'ID of user.'
          required: true
          example: '1'
      responses:
        200:
          description: 'User deleted.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        401:
          description: Unauthenticated.
        403:
          description: 'Permissions error.'
        404:
          description: 'User not found.'
      security:
        -
          sanctum: []
    patch:
      tags:
        - User
      summary: 'Partly update user data.'
      description: 'Only for admins.'
      operationId: 7f4b042564dfd3e52fc25d585c5e1da0
      parameters:
        -
          name: id
          in: path
          description: 'ID of user.'
          required: true
          example: '1'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchUserRequest'
      responses:
        200:
          description: 'User updated.'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Info message.'
                  entity:
                    $ref: '#/components/schemas/User'
                  old_entity:
                    $ref: '#/components/schemas/User'
                type: object
        401:
          description: Unauthenticated.
        403:
          description: 'Permissions error.'
        404:
          description: 'User not found.'
        422:
          description: 'Validation error.'
      security:
        -
          sanctum: []
  /user/info:
    get:
      tags:
        - User
      summary: 'Show info of loginned user.'
      operationId: 6d570e9f19903af4e05356943cbd8c88
      responses:
        200:
          description: 'Info about current user.'
          content:
            application/json:
              schema:
                properties:
                  entity:
                    $ref: '#/components/schemas/User'
                  roles:
                    type: array
                    items:
                      type: string
                      example: admin
                  groups:
                    type: array
                    items:
                      $ref: '#/components/schemas/Group'
                  links:
                    type: array
                    items:
                      $ref: '#/components/schemas/Link'
                type: object
        401:
          description: Unauthenticated.
        403:
          description: 'Permissions error.'
      security:
        -
          sanctum: []
components:
  schemas:
    MessageResponse:
      properties:
        message:
          type: string
          example: 'Info message'
      type: object
...
    Group:
      properties:
        id:
          type: integer
          example: '1'
        name:
          type: string
          example: 'Example group'
        description:
          type: string
          example: 'Example description'
        count:
          type: integer
          example: '0'
        created_at:
          type: string
          format: date-time
          example: '2023-09-27T06:16:30.000000Z'
        updated_at:
          type: string
          format: date-time
          example: '2023-09-27T06:16:30.000000Z'
      type: object
...
    Link:
      properties:
        id:
          type: integer
          example: '1'
        name:
          type: string
          example: 'Link example'
        description:
          type: string
          example: 'Example of link description.'
        origin:
          type: string
          example: 'https://example.url/11111'
        referral:
          type: string
          example: go1ZKwZNlg
        created_at:
          type: string
          format: date-time
          example: '2023-09-27T06:16:30.000000Z'
        updated_at:
          type: string
          format: date-time
          example: '2023-09-27T06:16:30.000000Z'
      type: object
...
    PatchUserRequest:
      properties:
        email:
          type: string
          example: example@mail.com
        password:
          type: string
          example: password
        name:
          type: string
          example: Example
        role_id:
          type: array
          items:
            type: integer
            example:
              - '1'
              - '2'
              - '3'
      type: object
    PutUserRequest:
      required:
        - email
        - password
        - name
      properties:
        email:
          type: string
          example: example@mail.com
        password:
          type: string
          example: password
        name:
          type: string
          example: Example
        role_id:
          type: array
          items:
            type: integer
            example: '1'
      type: object
    StoreUserRequest:
      required:
        - email
        - password
        - name
      properties:
        email:
          type: string
          example: example@mail.com
        password:
          type: string
          example: password
        name:
          type: string
          example: Example
      type: object
...
    User:
      properties:
        id:
          type: integer
          example: '1'
        name:
          type: string
          example: Example
        email:
          type: string
          example: example@mail.com
        email_verified_at:
          type: string
          format: date-time
          example: '2023-09-27T06:16:30.000000Z'
        created_at:
          type: string
          format: date-time
          example: '2023-09-27T06:16:30.000000Z'
        updated_at:
          type: string
          format: date-time
          example: '2023-09-27T06:16:30.000000Z'
      type: object
    UsersListPaginated:
      properties:
        current_page:
          type: integer
          example: '1'
        data:
          type: array
          items:
            $ref: '#/components/schemas/User'
        first_page_url:
          type: string
          example: 'http://links.com/user?page=1'
        from:
          type: integer
          example: '1'
        last_page:
          type: integer
          example: '2'
        last_page_url:
          type: string
          example: 'http://links.com/user?page=2'
        links:
          type: array
          items:
            properties:
              url:
                type: string
                example: 'null | http://links.com/user?page=2'
              label:
                type: string
                example: 'Next &raquo'
              active:
                type: boolean
                example: 'false'
            type: object
        next_page_url:
          type: string
          example: 'null | http://links.com/user?page=2'
        path:
          type: string
          example: 'http://links.com/user'
        per_page:
          type: integer
          example: '10'
        prev_page_url:
          type: string
          example: 'null | http://links.com/user?page=2'
        to:
          type: integer
          example: '10'
        total:
          type: integer
          example: '15'
      type: object
  securitySchemes:
    sanctum:
      type: apiKey
      description: 'Enter token in format (Bearer <token>)'
      name: Authorization
      in: header